Actualización del formulario de transacción financiera, de las vistas, adicion de nuevos reportes y eliminacino de auntiguos

+++ contabilidad/models.py
@@ -19,7 +19,7 @@ class Transaccion(models.Model):
     id = models.AutoField(primary_key=True)
     fecha = models.DateField()
     descripcion = models.CharField(max_length=255, null=True)
-    importe = models.FloatField()
+    importe = models.DecimalField(max_digits=10, decimal_places=2)  # Cambiado a DecimalField
     categoria = models.ForeignKey(Categoria, on_delete=models.CASCADE)
     cuenta = models.ForeignKey(Cuenta, on_delete=models.CASCADE)
